Added quote component to the dashboard. No problems atm
