+++ app/assets/stylesheets/application.scss
@@ -382,3 +382,49 @@ a:hover {
         background-size: cover;
         background-repeat: no-repeat;
     }
+
+
+// Quote Component
+
+.quote-box {
+    background: white;
+    width: 30rem;
+    margin: 10rem auto;
+    border-radius: 5px;
+    padding: 3rem;
+}
+
+.author {
+    display:flex;
+    justify-content: flex-end;
+}
+
+.quote-buttons {
+    display:flex;
+    justify-content: flex-end;
+}
+
+.new-quote{
+    background-color: #E6F2F2;
+    border: none;
+    padding: .7rem 1rem;
+    font-weight: bold;
+    border-radius: 2px;
+    outline: none;
+    cursor: pointer;
+
+}
+
+.new-quote:hover{
+    opacity: 0.7;
+}
+
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+body {
+    background: #E6F2F2;
+}
