+++ app/javascript/components/App.js
@@ -18,6 +18,7 @@ import Sidebar from "./components/Sidebar"
 import "./app.css"
 import DailyQuote from "./pages/DailyQuote"
 import ShowCard from "./components/ShowCard"
+import QuoteComp from "./components/QuoteComp"
 
 
 const App = (props) => {
@@ -92,7 +93,6 @@ const App = (props) => {
     <Route exact path="/*" element={<NotFound />} />
     <Route exact path="/workoutedit/:id"  element={<WorkoutEdit workouts = {workouts} {...props} updateWorkout={updateWorkout}/>} />
     <Route exact path="/workoutshow/:id"  element={<WorkoutShow workouts = {workouts} deleteWorkout = { deleteWorkout } {...props} />} />
-    <Route exact path="/dailyquote" element={<DailyQuote/>} />
     </Routes>
     <Footer />
   </BrowserRouter>
