+++ app/javascript/components/pages/Dashboard.js
@@ -1,6 +1,7 @@
 import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import ShowCard from '../components/ShowCard';
+import QuoteComp from '../components/QuoteComp';
 
 const Dashboard = (props) => {
   const navigate = useNavigate()
@@ -14,6 +15,7 @@ const Dashboard = (props) => {
     <>
     <div>Dashboard</div>
     <ShowCard/>
+    <QuoteComp/>
     </>
   )
 }
