--- app/javascript/components/pages/dailyquote.css
@@ -1,42 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    background: #E6F2F2;
-}
-
-#quote-box {
-    background: white;
-    width: 30rem;
-    margin: 10rem auto;
-    border-radius: 5px;
-    padding: 3rem;
-}
-
-#author {
-    display:flex;
-    justify-content: flex-end;
-}
-
-#buttons {
-    display:flex;
-    justify-content: flex-end;
-}
-
-#new-quote{
-    background-color: #E6F2F2;
-    border: none;
-    padding: .7rem 1rem;
-    font-weight: bold;
-    border-radius: 2px;
-    outline: none;
-    cursor: pointer;
-
-}
-
-#new-quote:hover{
-    opacity: 0.7;
-}
